+++ src/pages/people/outlets/color/palette/stories/Palette.stories.tsx
@@ -1,21 +1,26 @@
 import { StoryFn } from "@storybook/react";
 import { BrowserRouter } from "react-router-dom";
-import { Palette } from "./index";
+import { Palette } from "..";
 import { TokenContext } from "@context/TokenContext";
-import { presente } from "@src/mocks/design/tokensWithReference/presente";
+import { tokensWithReference } from "@src/mocks/design/tokensWithReference";
+import { props } from "./props";
 
 const story = {
-  components: [Palette],
+  components: Palette,
   title: "layouts/people/outlets/color/palette",
   parameters: {
     layout: "fullscreen",
   },
+  argTypes: props,
   decorators: [
-    (Story: StoryFn) => (
+    (
+      Story: StoryFn,
+      context: { args: { formType: keyof typeof tokensWithReference } }
+    ) => (
       <BrowserRouter>
         <TokenContext.Provider
           value={{
-            tokenWithRef: presente,
+            tokenWithRef: tokensWithReference[context.args.formType],
             loading: false,
             handleSubmit: () => {},
           }}
@@ -28,6 +33,10 @@ const story = {
 };
 
 const Default = () => <Palette />;
+
+Default.args = {
+  formType: "presente",
+};
 
 export default story;
