+++ src/pages/people/outlets/color/strokes/stories/Strokes.stories.tsx
@@ -1,21 +1,26 @@
 import { StoryFn } from "@storybook/react";
 import { BrowserRouter } from "react-router-dom";
-import { Surfaces } from "./index";
+import { Strokes } from "..";
 import { TokenContext } from "@context/TokenContext";
-import { presente } from "@src/mocks/design/tokensWithReference/presente";
+import { props } from "./props";
+import { tokensWithReference } from "@src/mocks/design/tokensWithReference";
 
 const story = {
-  components: [Surfaces],
-  title: "layouts/people/outlets/color/surfaces",
+  components: Strokes,
+  title: "layouts/people/outlets/color/strokes",
   parameters: {
     layout: "fullscreen",
   },
+  argTypes: props,
   decorators: [
-    (Story: StoryFn) => (
+    (
+      Story: StoryFn,
+      context: { args: { formType: keyof typeof tokensWithReference } }
+    ) => (
       <BrowserRouter>
         <TokenContext.Provider
           value={{
-            tokenWithRef: presente,
+            tokenWithRef: tokensWithReference[context.args.formType],
             loading: false,
             handleSubmit: () => {},
           }}
@@ -27,7 +32,11 @@ const story = {
   ],
 };
 
-const Default = () => <Surfaces />;
+const Default = () => <Strokes />;
+
+Default.args = {
+  formType: "presente",
+};
 
 export default story;
