+++ src/pages/people/outlets/color/surfaces/stories/Surfaces.stories.tsx
@@ -1,21 +1,26 @@
 import { StoryFn } from "@storybook/react";
 import { BrowserRouter } from "react-router-dom";
-import { Texts } from ".";
+import { Surfaces } from "..";
 import { TokenContext } from "@context/TokenContext";
-import { presente } from "@src/mocks/design/tokensWithReference/presente";
+import { tokensWithReference } from "@src/mocks/design/tokensWithReference";
+import { props } from "./props";
 
 const story = {
-  components: [Texts],
-  title: "layouts/people/outlets/color/texts",
+  components: Surfaces,
+  title: "layouts/people/outlets/color/surfaces",
   parameters: {
     layout: "fullscreen",
   },
+  argTypes: props,
   decorators: [
-    (Story: StoryFn) => (
+    (
+      Story: StoryFn,
+      context: { args: { formType: keyof typeof tokensWithReference } }
+    ) => (
       <BrowserRouter>
         <TokenContext.Provider
           value={{
-            tokenWithRef: presente,
+            tokenWithRef: tokensWithReference[context.args.formType],
             loading: false,
             handleSubmit: () => {},
           }}
@@ -27,7 +32,11 @@ const story = {
   ],
 };
 
-const Default = () => <Texts />;
+const Default = () => <Surfaces />;
+Default.args = {
+  formType: "presente",
+};
 
 export default story;
+
 export { Default };
